Tests: Updated How Conftest inputs data.
